requirements.txt
PyMuPDF==1.23.8
python-docx==1.1.0
pandas==2.1.4
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0
pytesseract==0.3.10
facenet-pytorch==2.5.3
spacy==3.7.2
openpyxl==3.1.2
python-pptx==0.6.23
gradio==4.13.0
google-generativeai==0.3.2
torch==2.1.2
torchvision==0.16.2